requirements.txt
Appium-Python-Client
pytest
opencv-python
numpy
